Made ResultSet returned in the mapper of withReturn default to the first row
This allows shorter / more readable mappers.

+++ src/main/java/nl/tijsbeek/database/databases/IStringDatabase.java
@@ -3,10 +3,10 @@ package nl.tijsbeek.database.databases;
 public interface IStringDatabase<Entity> extends IDatabase<Entity> {
 
     default Entity retrieveById(String id) {
-        return retrieveById(Integer.parseInt(id));
+        return retrieveById(Long.parseLong(id));
     }
 
     default Entity deleteById(String id) {
-        return deleteById(Integer.parseInt(id));
+        return deleteById(Long.parseLong(id));
     }
 }
